R36S: clarify init and audio host/token ABI boundaries

+++ src/libultra/audio/load.c
@@ -109,7 +109,7 @@ Acmd *alAdpcmPull(void *filter, s16 *outp, s32 outCount, s32 sampleOffset, Acmd 
          * Now fix up state info to reflect the loop start point
          */
         f->lastsam = f->loop.start &0xf;
-        f->memin = (s32) f->table->base + ADPCMFBYTES *
+        f->memin = (s32)(u32)(uintptr_t)f->table->base + ADPCMFBYTES *
             ((s32) (f->loop.start>>LFSAMPLES) + 1);
         f->sample = f->loop.start;
 
@@ -171,7 +171,7 @@ Acmd *alAdpcmPull(void *filter, s16 *outp, s32 outCount, s32 sampleOffset, Acmd 
      * overFlow is the number of bytes past the end
      * of the bitstream I try to generate
      */
-    overFlow = f->memin + nbytes - ((s32) f->table->base + f->table->len);
+    overFlow = f->memin + nbytes - ((s32)(u32)(uintptr_t)f->table->base + f->table->len);
     if (overFlow < 0)
         overFlow = 0;
     nOver = (overFlow/ADPCMFBYTES)<<LFSAMPLES;
@@ -277,7 +277,7 @@ Acmd *alRaw16Pull(void *filter, s16 *outp, s32 outCount, s32 sampleOffset, Acmd 
          */
         *outp += dramAlign;
         
-        f->memin = (s32) f->table->base + (f->loop.start<<1);
+        f->memin = (s32)(u32)(uintptr_t)f->table->base + (f->loop.start<<1);
         f->sample = f->loop.start;
         op = *outp;
         
@@ -338,7 +338,7 @@ Acmd *alRaw16Pull(void *filter, s16 *outp, s32 outCount, s32 sampleOffset, Acmd 
      */
 
     nbytes = outCount<<1;
-    overFlow = f->memin + nbytes - ((s32) f->table->base + f->table->len);
+    overFlow = f->memin + nbytes - ((s32)(u32)(uintptr_t)f->table->base + f->table->len);
     if (overFlow < 0)
         overFlow = 0;
     if (overFlow > nbytes)
@@ -408,7 +408,7 @@ alLoadParam(void *filter, s32 paramID, void *param)
                         (a->table->type == AL_ADPCM_WAVE) ? (void *)a->table->waveInfo.adpcmWave.book : NULL);
             }
 #endif
-            a->memin = (s32) a->table->base;
+            a->memin = (s32)(u32)(uintptr_t)a->table->base;
             a->sample = 0;
             switch (a->table->type){
                 case (AL_ADPCM_WAVE):
